Fixed imports

+++ script/Elevator.s.sol
@@ -1,14 +1,14 @@
 pragma solidity ^0.8.10;
 
-import {Script, console2} from "forge-std/Script.sol";
-import "../src/Elevator/ElevatorFactory.sol";
-import "../src/Elevator/ElevatorHack2.sol";
-import "../src/Ethernaut.sol";
+import { Script, console2 } from 'forge-std/Script.sol';
+import '../src/Elevator/ElevatorFactory.sol';
+import '../src/Elevator/ElevatorHack.sol';
+import '../src/Ethernaut.sol';
 
 contract ElevatorScript is Script {
     Ethernaut ethernaut;
 
-    function setUp() public {}
+    function setUp() public { }
 
     function run() public {
         vm.startBroadcast();
@@ -18,18 +18,17 @@ contract ElevatorScript is Script {
         address levelAddress = ethernaut.createLevelInstance(elevatorFactory);
         Elevator elevator = Elevator(payable(levelAddress));
 
-        console2.log("msg.sender is %s", msg.sender);
+        console2.log('msg.sender is %s', msg.sender);
 
         ElevatorHack elevatorHack = new ElevatorHack(levelAddress);
         elevatorHack.attack();
-        
+
         // msg.sender needs to be Building()
 
-
         bool top = elevator.top();
-        uint256 floor = elevator.floor();
-        console2.log("top: %s ", top);
-        console2.log("floor: %s ", floor);
+        uint floor = elevator.floor();
+        console2.log('top: %s ', top);
+        console2.log('floor: %s ', floor);
 
         bool levelSuccessfullyPassed = ethernaut.submitLevelInstance(payable(levelAddress));
         assert(levelSuccessfullyPassed);
